Te 6

See merge request !108

+++ trek-engine-client/src/app/app.component.ts
@@ -31,35 +31,49 @@ export class AppComponent implements OnInit {
 		let publicPath = ['/login', '/register', '/home', '/forgot-password'];
 		if(JSON.stringify(this._cookieService.getAll()) !== '{}'){
 			if(this._cookieService.get('authToken') && this._cookieService.get('authToken').length>0){
-				this.auth.validateToken(this._cookieService.get('authToken'))
-					.subscribe(validationResponse=>{
-						if(JSON.stringify(validationResponse) !== '{}'){
-							if(publicPath.includes(this.location.path()) || (this.location.path().indexOf('/change-password/token/') > -1) || (this.location.path().indexOf('/authorization/token/') > -1) ){
-								this._route.navigate(['/app']);
-							}else{
-								if(this.location && this.location.path() === ''){
+				if((this.location.path().indexOf('/register/validate?') > -1)){
+					let queryUrl = {};
+					if (location.search) location.search.substr(1).split("&").forEach(item => {let [k,v] = item.split("="); v = v && decodeURIComponent(v); (queryUrl[k] = queryUrl[k] || []).push(v)})
+					let navigateUrl = ['/register/validate',queryUrl['code'][0], queryUrl['loginType'][0], queryUrl['authuser'][0], queryUrl['prompt'][0], queryUrl['session_state'][0]]
+					this._route.navigate(navigateUrl);
+				}else{
+					this.auth.validateToken(this._cookieService.get('authToken'))
+						.subscribe(validationResponse=>{
+							if(JSON.stringify(validationResponse) !== '{}'){
+								if(publicPath.includes(this.location.path()) || (this.location.path().indexOf('/change-password/token/') > -1) || (this.location.path().indexOf('/authorization/token/') > -1) ){
 									this._route.navigate(['/app']);
 								}else{
-									this._route.navigate([this.location.path()]);
+									if(this.location && this.location.path() === ''){
+										this._route.navigate(['/app']);
+									}else{
+										this._route.navigate([this.location.path()]);
+									}
 								}
+							}else{
+								this._cookieService.removeAll();
+								window.location.href = environment.webUrl;
 							}
-						}else{
+						}, error =>{
 							this._cookieService.removeAll();
 							window.location.href = environment.webUrl;
-						}
-					}, error =>{
-						this._cookieService.removeAll();
-						window.location.href = environment.webUrl;
-					});
+						});
+				}
 			}else{
 				this._cookieService.removeAll();
 				this._route.navigate(['/home']);
 			}
 		}else{
-			if( publicPath.includes(this.location.path()) || (this.location.path().indexOf('/change-password/token/') > -1) || (this.location.path().indexOf('/authorization/token/') > -1) ){
+			if( publicPath.includes(this.location.path()) || (this.location.path().indexOf('/change-password/token/') > -1) || (this.location.path().indexOf('/authorization/token/') > -1) || (this.location.path().indexOf('/register/code/') > -1) ){
 				this._route.navigate([this.location.path()]);
 			}else{
-				this._route.navigate(['/home']);
+				if((this.location.path().indexOf('/register/validate?') > -1)){
+					let queryUrl = {};
+					if (location.search) location.search.substr(1).split("&").forEach(item => {let [k,v] = item.split("="); v = v && decodeURIComponent(v); (queryUrl[k] = queryUrl[k] || []).push(v)})
+					let navigateUrl = ['/register/validate',queryUrl['code'][0], queryUrl['loginType'][0], queryUrl['authuser'][0], queryUrl['prompt'][0], queryUrl['session_state'][0]]
+					this._route.navigate(navigateUrl);
+				}else{
+					this._route.navigate(['/home']);
+				}
 			}
 		}
 	}
